SMTPADDONS-114 - Issue fix for Beta Version

+++ Postman/PostmanInputSanitizer.php
@@ -104,7 +104,11 @@ if ( ! class_exists( 'PostmanInputSanitizer' ) ) {
 				$new_input = apply_filters( 'post_smtp_sanitize', $new_input, $input, $this );
 				delete_transient( 'sendpulse_token' );
 
-				if ( $new_input [ PostmanOptions::CLIENT_ID ] != $this->options->getClientId() || $new_input [ PostmanOptions::CLIENT_SECRET ] != $this->options->getClientSecret() || $new_input [ PostmanOptions::HOSTNAME ] != $this->options->getHostname() ) {
+				if (
+					( isset( $new_input[ PostmanOptions::CLIENT_ID ] ) && $new_input[ PostmanOptions::CLIENT_ID ] != $this->options->getClientId() ) ||
+					( isset( $new_input[ PostmanOptions::CLIENT_SECRET ] ) && $new_input[ PostmanOptions::CLIENT_SECRET ] != $this->options->getClientSecret() ) ||
+					( isset( $new_input[ PostmanOptions::HOSTNAME ] ) && $new_input[ PostmanOptions::HOSTNAME ] != $this->options->getHostname() )
+				){
 					$this->logger->debug( 'Recognized new Client ID' );
 					// the user entered a new client id and we should destroy the stored auth token
 					delete_option( PostmanOAuthToken::OPTIONS_NAME );
